Review and create - extract summary tables into separate files
